+++ src/__tests__/helpers/testHelpers.ts
@@ -2,7 +2,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2019-Present Datadog, Inc.
 
-import { Stats, Report, Compilation, Compiler } from '../../types';
+import { BundlerStats, Stats, Report, Compilation, Compiler } from '../../types';
 
 export const mockStats = ({
     toJson: jest.fn(() => ({
@@ -27,11 +27,23 @@ export const mockStats = ({
     },
 } as unknown) as Stats;
 
+export const mockBundler: BundlerStats = {
+    webpack: mockStats,
+    esbuild: {
+        warnings: [],
+        errors: [],
+        entrypoints: [],
+        duration: 0,
+        inputs: {},
+        outputs: {},
+    },
+};
+
 export const mockReport = {
     timings: {
-        tapables: {},
-        loaders: {},
-        modules: {},
+        tapables: new Map(),
+        loaders: new Map(),
+        modules: new Map(),
     },
     dependencies: {},
 } as Report;
