+++ src/hooks/datadog/__tests__/aggregator.test.ts
@@ -3,175 +3,13 @@
 // Copyright 2019-Present Datadog, Inc.
 
 import { mockReport, mockStats } from '../../../__tests__/helpers/testHelpers';
-import path from 'path';
-import { getModules, getIndexed, getEntries, getChunks, getAssets } from '../metrics/webpack';
-import { StatsJson } from '../../../types';
-import { Metric } from '../types';
-
-const exec = require('util').promisify(require('child_process').exec);
-
-const PROJECTS_ROOT = path.join(__dirname, '../../../__tests__/mocks/projects');
 
 describe('Aggregator', () => {
-    beforeAll(async () => {
-        await exec(`yarn build`);
-    }, 20000);
-
-    for (const version of [4, 5]) {
-        describe(`Webpack ${version}`, () => {
-            let statsJson: StatsJson;
-            const WEBPACK_ROOT = path.join(PROJECTS_ROOT, `./webpack${version}`);
-            const OUTPUT = path.join(WEBPACK_ROOT, './webpack-profile-debug/');
-
-            beforeAll(async () => {
-                const output = await exec(`yarn workspace webpack${version} build`);
-
-                // eslint-disable-next-line no-console
-                console.log(`Build ${version} :`, output.stderr);
-
-                statsJson = require(path.join(OUTPUT, './stats.json'));
-            }, 20000);
-
-            test('It should aggregate metrics without throwing.', () => {
-                const { getMetrics } = require('../aggregator');
-                const opts = { context: '', filters: [], tags: [] };
-                expect(() => {
-                    getMetrics(opts, mockReport, mockStats);
-                }).not.toThrow();
-            });
-
-            describe('Modules', () => {
-                let metrics: Metric[];
-
-                beforeAll(() => {
-                    const indexed = getIndexed(statsJson, WEBPACK_ROOT);
-                    metrics = getModules(indexed, WEBPACK_ROOT);
-                });
-
-                test('It should give module metrics.', () => {
-                    expect(metrics.length).not.toBe(0);
-                });
-
-                test(`It should filter out webpack's modules.`, () => {
-                    expect(
-                        metrics.find((m) => {
-                            return m.tags.find((t) => /^moduleName:webpack\/runtime/.test(t));
-                        })
-                    ).toBeUndefined();
-                });
-
-                test(`It should add tags about the entry and the chunk.`, () => {
-                    for (const metric of metrics) {
-                        expect(metric.tags).toContain('entryName:yolo');
-                        expect(metric.tags).toContain('entryName:cheesecake');
-                        expect(metric.tags).toContain('chunkName:yolo');
-                        expect(metric.tags).toContain('chunkName:cheesecake');
-                    }
-                });
-
-                test('It should have 3 metrics per module.', () => {
-                    const modules = [
-                        './src/file0000.js',
-                        './src/file0001.js',
-                        './workspaces/app/file0000.js',
-                        './workspaces/app/file0001.js',
-                    ];
-
-                    for (const module of modules) {
-                        const modulesMetrics = metrics.filter((m) =>
-                            m.tags.includes(`moduleName:${module}`)
-                        );
-                        expect(modulesMetrics.length).toBe(3);
-                    }
-                });
-            });
-
-            describe('Entries', () => {
-                let metrics: Metric[];
-
-                beforeAll(() => {
-                    const indexed = getIndexed(statsJson, WEBPACK_ROOT);
-                    metrics = getEntries(statsJson, indexed);
-                });
-
-                test('It should give entries metrics.', () => {
-                    expect(metrics.length).not.toBe(0);
-                });
-
-                test('It should give 4 metrics per entry.', () => {
-                    const entries = ['yolo', 'cheesecake'];
-
-                    for (const entry of entries) {
-                        const entriesMetrics = metrics.filter((m) =>
-                            m.tags.includes(`entryName:${entry}`)
-                        );
-                        expect(entriesMetrics.length).toBe(4);
-                    }
-                });
-            });
-
-            describe('Chunks', () => {
-                let metrics: Metric[];
-
-                beforeAll(() => {
-                    const indexed = getIndexed(statsJson, WEBPACK_ROOT);
-                    metrics = getChunks(statsJson, indexed);
-                });
-
-                test('It should give chunks metrics.', () => {
-                    expect(metrics.length).not.toBe(0);
-                });
-
-                test('It should give 2 metrics per chunk.', () => {
-                    const chunks = ['yolo', 'cheesecake'];
-
-                    for (const chunk of chunks) {
-                        const chunksMetrics = metrics.filter((m) =>
-                            m.tags.includes(`chunkName:${chunk}`)
-                        );
-                        expect(chunksMetrics.length).toBe(2);
-                    }
-                });
-
-                test(`It should add tags about the entry.`, () => {
-                    for (const metric of metrics) {
-                        expect(metric.tags.join(',')).toMatch(/entryName:(yolo|cheesecake)/);
-                    }
-                });
-            });
-
-            describe('Assets', () => {
-                let metrics: Metric[];
-
-                beforeAll(() => {
-                    const indexed = getIndexed(statsJson, WEBPACK_ROOT);
-                    metrics = getAssets(statsJson, indexed);
-                });
-
-                test('It should give assets metrics.', () => {
-                    expect(metrics.length).not.toBe(0);
-                });
-
-                test('It should give 1 metric per asset.', () => {
-                    const assets = ['yolo.js', 'cheesecake.js'];
-
-                    for (const asset of assets) {
-                        const assetsMetrics = metrics.filter((m) =>
-                            m.tags.includes(`assetName:${asset}`)
-                        );
-                        expect(assetsMetrics.length).toBe(1);
-                    }
-                });
-
-                test(`It should add tags about the entry and the chunk.`, () => {
-                    for (const metric of metrics) {
-                        expect(metric.tags).toContain('entryName:yolo');
-                        expect(metric.tags).toContain('entryName:cheesecake');
-                        expect(metric.tags).toContain('chunkName:yolo');
-                        expect(metric.tags).toContain('chunkName:cheesecake');
-                    }
-                });
-            });
-        });
-    }
+    test('It should aggregate metrics without throwing.', () => {
+        const { getMetrics } = require('../aggregator');
+        const opts = { context: '', filters: [], tags: [] };
+        expect(() => {
+            getMetrics(opts, mockReport, mockStats);
+        }).not.toThrow();
+    });
 });
